+++ src/api/java/mekanism/api/datagen/recipe/builder/PressurizedReactionRecipeBuilder.java
@@ -19,7 +19,7 @@ public class PressurizedReactionRecipeBuilder extends MekanismRecipeBuilder<Pres
     private final ItemStackIngredient inputSolid;
     private final FluidStackIngredient inputFluid;
     private final ChemicalStackIngredient inputChemical;
-    private long energyRequired = 0;
+    private int energyRequired = 0;
     private final int duration;
     @Nullable
     private final ItemStackTemplate outputItem;
@@ -106,7 +106,7 @@ public class PressurizedReactionRecipeBuilder extends MekanismRecipeBuilder<Pres
      *
      * @param energyRequired How much "extra" energy this recipe requires, compared to the base energy requirements of the machine performing the recipe.
      */
-    public PressurizedReactionRecipeBuilder energyRequired(long energyRequired) {
+    public PressurizedReactionRecipeBuilder energyRequired(int energyRequired) {
         if (energyRequired < 0) {
             throw new IllegalArgumentException("This reaction recipe must have a positive or zero energy requirement.");
         }
