+++ src/main/java/mekanism/common/attachments/containers/energy/ComponentBackedNoClampEnergyContainer.java
@@ -1,5 +1,6 @@
 package mekanism.common.attachments.containers.energy;
 
+import java.util.function.IntSupplier;
 import java.util.function.LongSupplier;
 import java.util.function.Predicate;
 import mekanism.api.AutomationType;
@@ -11,7 +12,7 @@ import org.jetbrains.annotations.NotNull;
 public class ComponentBackedNoClampEnergyContainer extends ComponentBackedEnergyContainer {
 
     public ComponentBackedNoClampEnergyContainer(ItemAccess attachedAccess, int containerIndex, Predicate<@NotNull AutomationType> canExtract,
-          Predicate<@NotNull AutomationType> canInsert, LongSupplier rate, LongSupplier maxEnergy) {
+          Predicate<@NotNull AutomationType> canInsert, IntSupplier rate, LongSupplier maxEnergy) {
         super(attachedAccess, containerIndex, canExtract, canInsert, rate, maxEnergy);
     }
